add -force unison option to prevent unexpected behavior in most cases
With this change the unison container will first make sure the UNISON_DIR has an exact copy of UNISON_HOST_DIR
This prevents problems, when the unison container has been stopped and removed, and some host files have been removed.
In this case these files will still exist in the UNISON_DIR volume, so unison would copy them back to the host.

The side effect is that any changes that are just in UNISON_DIR will be lost when the unison container starts up.